requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-cors-middleware==1.4.0
django-crispy-forms==1.8.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
httplib2==0.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
Pygments==2.4.2
PyJWT==1.7.1
python-slugify==4.0.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.0
urllib3==1.25.6
